requirements.txt
streamlit
openai
python-dotenv
pandas
numpy
reportlab
plotly